feat: decrease number of layers and connect with mongodb

+++ src/domain/use-cases/project/delete-project.ts
@@ -9,6 +9,15 @@ export class DeleteProject implements DeleteProjectUseCase {
   }
 
   async execute(id: string): Promise<void> {
-    await this.projectRepository.deleteProject(id);
+    try {
+      const projectExists = await this.projectRepository.getProject(id);
+      if (!projectExists) {
+        throw new Error("Projeto não encontrado");
+      }
+      await this.projectRepository.deleteProject(id);
+    } catch (error) {
+      console.log(error);
+      throw new Error("Houve um erro ao deletar o projeto");
+    }
   }
 }
